중간 수정 사항

+++ mydata/src/main/java/kr/ac/kopo/mydata/dto/BankStatementDTO.java
@@ -1,5 +1,6 @@
 package kr.ac.kopo.mydata.dto;
 
+import java.math.BigDecimal;
 import java.util.Date;
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -9,7 +10,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class BankStatementDTO {
-  private Integer transactionNo;
+  private BigDecimal transactionNo;
   private String accountNo;
   private Integer transactionAmount;
   private Date transactionDate;
